% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermatrix.R
\name{colour_correlation}
\alias{colour_correlation}
\title{Colour Correlation}
\usage{
colour_correlation(data, mapping, method = "p", use = "pairwise", ...)
}
\description{
This is an internal function of \code{\link{plot_scattermatrix}} and
it is not supposed to be called directly.
}
\details{
Function to return coloured correlations
}
