% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermatrix.R
\name{make_pairs}
\alias{make_pairs}
\title{Make Pairs for Scatter Matrix}
\usage{
make_pairs(data, ...)
}
\arguments{
\item{data}{Dataframe or tibble}

\item{...}{Variables to select}
}
\value{
Returns a reshaped dataframe.
}
\description{
Make Pairs for Scatter Matrix
}
\examples{

make_pairs(iris)
make_pairs(iris, -Species)
make_pairs(mtcars, cyl, mpg, hp)

}
