% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a Heatmap}
\usage{
plot_heatmap(
  data,
  order_rows = FALSE,
  order_cols = FALSE,
  title = "Heatmap",
  scalefun = scales::rescale
)
}
\arguments{
\item{data}{Tibble or dataframe. Data to plot.}

\item{order_rows}{Logical. Reorder rows?}

\item{order_cols}{Logical. Reorder rows?}

\item{title}{String. Title of the heatmap.}
}
\value{
Returns a ggplot geom_tile heatmap
}
\description{
Plot a Heatmap
}
\examples{
plot_heatmap(iris)
plot_heatmap(mtcars, order_rows = TRUE, order_cols = TRUE)
plot_heatmap(mtcars)
plot_heatmap(mtcars, order_rows = TRUE)
plot_heatmap(mtcars, order_cols = TRUE)
plot_heatmap(mtcars, order_rows = TRUE, order_cols = TRUE)

}
