% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{plot_pareto_histogram}
\alias{plot_pareto_histogram}
\title{Build a Pareto Chart from an Histogram}
\usage{
plot_pareto_histogram(
  to_plot,
  variable,
  binwidth = NULL,
  xlab = variable,
  ylab = "Percentage",
  title = paste("Pareto Histogram -", quo_name(xlab)),
  ylab2 = "\% cumulate",
  hline = 0.8,
  bar.color = "gray70",
  bar.fill = "steelblue",
  bar.alpha = 0.6,
  line.color = "forestgreen",
  line.size = 1,
  hline.col = "coral2",
  boundary = 0
)
}
\arguments{
\item{to_plot}{tibble of interest.}

\item{xlab}{title of x axis, default: "items".}

\item{ylab}{title of left y axis, default: "".}

\item{title}{title of the chart, default: "Pareto Chart".}

\item{ylab2}{title of right y axis, default: "\% cumulate".}

\item{hline}{yintercept of horizontal line, default: 0.8.}

\item{bar.color}{colour of bars, default: "royalblue4".}

\item{bar.fill}{fill of bars, default: "royalblue".}

\item{bar.alpha}{transparency of bars, default: 0.4.}

\item{line.color}{color of cumulate line, default: "green".}

\item{line.size}{size of cumulate line, default: 1.}

\item{hline.col}{color of horizontal line, default: "orange"}

\item{modality}{column you want to show the frequency.}

\item{frequency}{frequency of modality.}

\item{group_over}{number of max modalities to show. Sum the others. Default: NULL}

\item{others_name}{name to assign to modalities over group_over. Default: "other"}
}
\value{
Returns a Pareto Chart, ggplot2 object
}
\description{
The function builds a Pareto Chart by summing up
the bars of an histogram throught a freqpoly
}
\examples{
test <- data.frame(x = rnorm(1100, 10, 3), g = letters[1:5])
plot_pareto_histogram(test, x)
plot_pareto_histogram(data.frame(x = c(1,1,10)), x)
plot_pareto_histogram_facet(test, x, g)

}
