% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{plot_pareto_histogram_facet}
\alias{plot_pareto_histogram_facet}
\title{Build a Pareto Chart from an Histogram with facets}
\usage{
plot_pareto_histogram_facet(
  to_plot,
  variable,
  facet = NULL,
  binwidth = NULL,
  xlab = variable,
  ylab = "Percentage",
  title = paste("Pareto Histogram -", quo_name(xlab)),
  hline = 0.8,
  bar.color = "gray70",
  bar.fill = "steelblue",
  bar.alpha = 0.6,
  line.color = "forestgreen",
  line.size = 1,
  hline.col = "coral2",
  boundary = 0
)
}
\arguments{
\item{facet}{variable to use for faceting}
}
\value{
Returns a Pareto Chart, ggplot2 object
}
\description{
The function builds a Pareto Chart by summing up
the bars of an histogram throught a freqpoly
}
\examples{
test <- data.frame(x = rnorm(1000), g = c("a","a", "b", "c"))
plot_pareto_histogram_facet(test, x, g)
test <- data.frame(x = rnorm(1000), g = letters[1:5])
plot_pareto_histogram_facet(test, x, g)

}
