% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermatrix.R
\name{plot_scattermatrix}
\alias{plot_scattermatrix}
\title{Plot a ScatterMatrix}
\usage{
plot_scattermatrix(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
ggmatrix object that if called, will print
}
\description{
Plot a ScatterMatrix
}
\examples{

plot_scattermatrix(iris)
plot_scattermatrix(mtcars)

}
