% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plot_silhouette}
\alias{plot_silhouette}
\title{Plot Silhouette}
\usage{
plot_silhouette(
  v_clusters,
  mx_dist,
  title = "Silhouette Graph - Check for cluster cohesion",
  threshold = 1
)
}
\arguments{
\item{v_clusters}{An integer vector with k different integer cluster codes or a list with such an x$clustering component. See ?cluster::silhouette}

\item{mx_dist}{A dissimilarity object inheriting from class dist. See ?cluster::silhouette}

\item{title}{String. Custom title.}

\item{threshold}{Numeric. Threshold over which an item is considered to be fitting its own cluster.}
}
\value{
A ggplot object with silhouette plot.
}
\description{
Returns a silhouette plot for cluster analysis. It helps to check for cluster cohesion.
}
\examples{

kcl <- kmeans(mtcars, centers = 5, nstart = 50)
dis <- dist(mtcars)^2
plot_silhouette(v_clusters = kcl$cluster, mx_dist = dis, threshold = 0.3)

}
\references{
Wiki link: https://en.wikipedia.org/wiki/Silhouette_(clustering)
}
