% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattermatrix.R
\name{points_and_smooth}
\alias{points_and_smooth}
\title{Points and Smooth}
\usage{
points_and_smooth(data, mapping, method = "loess", ...)
}
\description{
This is an internal function of \code{\link{plot_scattermatrix}} and
it is not supposed to be called directly.
}
\details{
Function to return points and geom_smooth
}
