% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Rescale between 0 and 1 with min-max}
\usage{
rescale(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector to be rescaled}
}
\value{
Returns rescaled x
}
\description{
Rescale between 0 and 1 with min-max
}
\examples{

rescale(c(1, 4, 5, 2, 9.1))

}
