% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall}
\usage{
plot_waterfall(
  data,
  starting_point = 0,
  label_starting_point = "starting_point",
  title = "Waterfall",
  subtitle = NULL
)
}
\arguments{
\item{data}{Tibble. See details and examples}

\item{starting_point}{Numeric. First column height.}

\item{label_starting_point}{String. First column label.}

\item{title}{String. Title of the plot}

\item{subtitle}{String. Subtitle of the plot.}
}
\value{
Returns a waterfall plot.
}
\description{
Plot waterfall
}
\details{
data must be a tibble with 3 columns:
\itemize{
\item group: for every group you'llhave an intermediate column
\item label: name of increase or decrease
\item value: increase or decrease value
}
}
\examples{

aa <- tibble::tribble(
  ~ group  , ~ label      , ~ value,
  "season1", "loss"       ,     -20,
  "season1", "new"        ,      30,
  "season1", "gain"       ,      40,
  "season2", "loss"       ,     -15,
  "season2", "gain"       ,      25,
  "season3", "loss"       ,     -10,
  "season3", "disappeared",     -15,
  "season3", "gain"       ,       5

)

starting_point <- 100

plot_waterfall(aa, starting_point)


}
